requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]>=4.3.0
certifi
scikit-learn
dill
PyYAML
mlflow
dagshub
fastapi
uvicorn
python-multipart
#-e.